refactor : 로직을 사용하는 구현 계층 변경
